Refactor TimerDataTest to make it easy to add new tests (#4108)
* Refactor TimerDataTest to make it easy to add new tests

This PR is just a refactor of TimerDataTest, so it's easy to add new
test. The next PR will add a test for GetTimers() both when timers are
inserted in-order and not-in-order. A following PR will include a new
method to get timers in an optimized way with its corresponding test.

While cleaning this test I found that some query is assuming timers to
be in ordered so I added a TODO. This might be one of the causes that
arrow keys wasn't working in GpuSubmissionTrack.

Bug: http://b/242971209

* Fixup: comment

* Fixup: Consistent use of static