Fix GetTimers by returning all visible timers
In this PR we are making GetTimers from TimerData to return all
visible timers. As far as I know, this method was implemented but
never used (http://b/204173236 - correct me if I am wrong Valeriy).

Since this method will be used to draw timers, it should include the
partially visible as well, similar to what happens with
ScopeTreeTimerData.

In addition, I'm adding tests to the method.

Test: UnitTest
Bug: http://b/204173236