Add orbit_base::NotFoundOr<T>
This is used as a return type for search operations that can fail. This
used to be done with ErrorMessageOr and an error when not found.
NotFoundOr is necessary to differentiate between real errors and the
failure of a search. The orbit_base::NotFound type contains a message
field, which is used to explain context why the search failed.

Test: Unit tests
Bug: http://b/231455031