Use NotFoundOr for GetDebugInfoFile
This changes the search for a symbols file on the instance to
differentiate between an actual error occurring and the symbols not
being found. This changes all the parts that are involved up to
OrbitApp. The communication protos do not need to be changed, because
the status is transmitted via the internal gRPC status class. This will
later be used to send different metrics, whether this was found or an
error occurred.

Test: Partially unit tests; manual test: clean cache, start Orbit, check
that symbol downloading still works.
Bug: http://b/231455031