Add tests to the GetNextPixel utility method (#4134)
* Add tests to the GetNextPixel utility method

In this PR we are adding tests to GetNextPixelBoundaryTimeNs() method.

The test showed that the method was not working fully optimal when the
resolution was not a divisor of the number of visible pixel, because it
was visiting twice each pixel (the max ns of the left pixel and the
min ns of the right pixel). I fixed it by making a ceiling instead. Now,
the function to get the min_timestamp for a given pixel return a
timestamp that will be assigned to the correct pixel.

Bug: http://b/243628546
Test: Unit Test. Compared visually pre-post state of a loaded capture
  - Pre: http://screenshot/7VWPTMz4SCqLsyf
  - Post: http://screenshot/7MDRgz6ZaxzoAQy

* Fixup: Anton comment + test

* Fixup: Anton's comments