Implement optimizated query in TimerData
In this PR we are implementing the optimized query in TimerData,
similarly to what we did in ThreadTracks (ScopeTreeTimerData). The query
was introduced in https://github.com/google/orbit/pull/4107, but not
implemented.

As it says in TimerDataInterface, GetTimersAtDepthDiscretized avoids
returning two timers that map to the same pixel in the screen, so is
especially useful when there are many timers in the screen (zooming-out
for example).

The implementation is way similar to ScopeTreeTimerData's one, but it
needs to iterate through blocks in the TimerChain because the data
structure doesn't provide O(1) access (explained in http://b/242970806).

We include a method in TimerChain to get the LowerBound inside a Block),
and we added unit tests in the query. In a following step, this method
will be used to simplify UpdatePrimitives in SchedulerTrack.

Bug: http://b/242971209.
Test: Unit Tests.