Parse PDB's public symbols using LLVM (#4131)
In a PDB file, public symbols are not a real subset of
the module info stream's symbols. There are symbols for
functions (like RtUserThreadStart) that are only present
in the public symbols.

This change adds support for reading public symbols
using LLVM.

Unlike DIA, LLVM is not extrapolating the size information
for the public symbols for us. Therefore, we try to look-up
the missing size information in the section contribution
stream. As a fallback, we also try to deduce the
size information from the distance to the next symbol.

Test: Run test + Check on a couple of Windows dlls.
Bug: http://b/221968448