Symbol loading outcome (#4140)
* Add orbit_base::NotFoundOr<T>

This is used as a return type for search operations that can fail. This
used to be done with ErrorMessageOr and an error when not found.
NotFoundOr is necessary to differentiate between real errors and the
failure of a search. The orbit_base::NotFound type contains a message
field, which is used to explain context why the search failed.

Test: Unit tests
Bug: http://b/231455031

* Add SymbolLoadingOutcome

SymbolLoadingOutcome expresses the result of a symbol loading task that
can fail with an error, get canceled, result in a "NotFound" outcome, or
success. This will be used as a return type to various symbol loading
operations. The `SuccessType` outcome contains a file path and 2 fields
that contain meta data about the search. These will be used for metrics
(compare SymbolLoadingMetric and/or orbit_log_event.proto)

Test: Unit Test
Bug: http://b/231455031