Make module's display name consistent
Elf files have an entry for the "soname", which might be different
from the filename, in case of semantic versioning.
E.g. for libc, we have "libc.so.6" as "soname" and "libc-2.24.so"
as filename.
For ElfFiles, we report the "soname" as "ModuleData::name".

Within Orbit, we used both, the filename and the soname accross
the codebase in various places.

This change unifies this now, to always use the filename. We
go with the filename for simplicity, as we do not always have
a module at hand in the places.
In practice is should not matter, and should be equally clear
to the user.

Test: Take a capture and inspect all places where we have module
      names.
Bug: http://b/222283082