Make DataView::GetToolTip return DataView::GetValue by default
There are plenty of cases where the content of a cell of a `DataView` doesn't
fit in the width of the column. In some cases, even expanding the column is not
enough to reveal the full content, as the `DataView` itself is not wide enough.
We enable tooltips everywhere by making it show the content of the cell by
default.

Test: Take a capture, verify tooltips in all `DataView`s.