Change logic to merge RUNTIME_FUNCTIONs into function address ranges
The current logic didn't consider the following from
https://docs.microsoft.com/en-us/cpp/build/exception-handling-x64#chained-unwind-info-structures:
"Chained info [...] can be used for noncontiguous code segments. By using
chained info, you can reduce the size of the required unwind information,
because you do not have to duplicate the unwind codes array from the primary
unwind info."

It only considered the more common case "You can also use chained info to group
volatile register saves. The compiler may delay saving some volatile registers
until it is outside of the function entry prolog."

In other words, it didn't consider that `RUNTIME_FUNCTION`s can be chained with
a primary `RUNTIME_FUNCTION` that actually belongs to another function and that
came way earlier, as a means of reusing unwind information from that function.
This is very rare as we only observed it for a handful of `RUNTIME_FUNCTION`s in
one large complex game.

The new logic to merge a group of `RUNTIME_FUNCTION`s into a single function is
that the `RUNTIME_FUNCTION`s must be adjacent (i.e., no "holes" allowed), and
that they must have the same primary `RUNTIME_FUNCTION`. But we don't require
the primary `RUNTIME_FUNCTION` to belong to the group, to support this new
corner case.

Test:
- Existing unit tests of `LoadDebugSymbols`, that uses this function.
- With unit tests that will come in a following change regarding loading these
  `UnwindRange`s as symbols.
- As part of the larger symbol loading fallback prototype.