The ThreadStateSlice proto now contains two additional fields, a callstack_id (that does not yet get set at all), and a status, that can be either of:
no callstack is and will be attached to this slice
a callstack will be attached to this slice; wait for it
there is a callstack attached (callstack id contains data)
With this change, we allow ThreadStateManager to actually specify that we are waiting for a callstack to be attached later (in the ProducerEventProcessor).
By now, the manager will not yet get called (which will happen in the next PR).

Test: Unit test
Bug: http://b/235554760