Add ElfFile::LoadEhOrDebugFrameEntriesAsSymbols
This is the counterpart of `CoffFile::LoadExceptionTableEntriesAsSymbols`
(https://github.com/google/orbit/pull/4178).
It uses `.debug_frame` if available and falls back to `.eh_frame`, which is
usually the only one present.
Again, it assigns an arbitrary name to the resulting function symbols.

We need some ugly and specific handling of `.eh_frame` entries as returned by
LLVM because of a bug only fixed in LLVM 13.

Bug: http://b/245680119

Test:
- Unit tests.
- As part of the larger symbol loading fallback prototype.