Add ObjectFile::LoadDynamicLinkingSymbolsAndUnwindRangesAsSymbols
The choice here is to provide a single virtual method in `ObjectFile`. This is
instead of providing a virtual method for loading dynamic linking symbols and a
virtual method for loading unwind information ranges, and then placing the logic
for merging these function at a higher level (e.g., outside of `ObjectFile`, or
in a non-virtual method of `ObjectFile`). The reason is that the logic to merge
the functions from these two sources is specific to the file type, as shown by
the differences in implementation between `ElfFileImpl` and `CoffFileImpl`, with
`CoffFileImpl::LoadDynamicLinkingSymbolsAndUnwindRangesAsSymbols` being more
complicated because of missing symbol sizes from the Export Table.

Also, `CoffFileImpl::LoadSymbolsFromExportTableInternal` and
`CoffFileImpl::LoadExceptionTableEntriesAsSymbolsInternal` are added to avoid
reading the `RUNTIME_FUNCTION`s twice (with `GetUnwindRanges()`.

Bug: http://b/245680119

Test:
- Unit tests.
- As part of the larger symbol loading fallback prototype.