Change orbit_ggp::Client to support not found
We change the orbit_ggp::Client to only support query one module while
calling GetSymbolDownloadInfoAsync to get the symbol download url.
If failed to get the symbol download url, GetSymbolDownloadInfoAsync
then checks the stderr of the ggp call to see whether it is a not found case
or some other error cases, and returns NotFound or an error message
accordingly.

Bug: http://b/245920841
Test: Unit tests