Add unit tests for unwinding of thread state slice callstacks
With #4182 we introduced processing tracepoint events with
stack data in uprobes unwinding visitor. For those events,
all our assumptions/requirements should be the same, so we
want to test the same set for thos.

This change split the existing unit tests for the unwinding visitor
into one part that is related to raw stack unwinding (aka dwarf)
and into another part (file) for the frame-pointer/callchain
patching.

For the DWARF cases, where we have implemented the tracepoint
processing, we made the test itself a type-parametrizied test.
The type parameter contains of the input PerfEvent and the output
proto.
Creation of events/protos as well as expecting different listener
calls, can not be done easily using type-parameters, so those
are handled using a compile-time switch (if constexpr(...))
for the different types.

Test: Run unit-test
Bug: http://b/244403766