Wait for dialog to be closed in E2E tests
In this PR we are waiting for dialogs to be closed in
orbit_capture_repeatedly and orbit_symbol_locations, so Alt+F4 close
and not the dialogs.

For capture_repeatedly we are waiting for the capture option button to
be enabled again and for symbol_locations we are waiting for the button
to be None.

Test: Run tests locally, will trigger E2E test in the signed build.