Merge thread state slice callstacks in ProducerEventProcessor
From the SwitchesStatesNamesVisitor and the UprobesUnwindingVisitor
we get two different events, 1) a ThreadStateSlice and 2) a
callstack.
The ThreadStateSlice indicates that it is waiting for a callstack.

Within the ProducerEventProcessor, we now
1) Intern the callstack
2) Assign the interned callstack id to the matching thread state
   slice.

Note that we can assume to see the callstacks always before the
thread state slices, as all callstacks correspond to the slice's
begin tracepoint.

Test: Unit Test
Bug: http://b/235554760