Add StructuredDebugDirectorySymbolProvider (#4193)
This adds the class StructuredDebugDirectorySymbolProvider to
orbit_symbol_providers and uses it when searching for symbols on the
local machine. This class will also be used in the future to search in
structured debug directories on the instance.

This finding in structured debug directories was (and still is) part of
SymbolHelper::FindSymbolsFileLocally. The signature of this method is
not changed in this PR, to keep this change fairly contained.

The main unit tests are moved into the new
StructuredDebugDirectorySymbolProviderTest. SymbolHelper keeps one test
to verify that FindSymbolsFileLocally still uses the structured debug
directories

Test: Unit test
Bug: http://b/245879461