Add FunctionsDataView::RemoveFunctionsOfModule
This will be used in the context of the "fallback symbols": if these are loaded,
but "Load Symbols" is selected and does now find symbols, the "fallback symbols"
need to be removed, as they now get completely replaced by the actual symbols.

Also simplify `FunctionsDataView::AddFunctions` because `OnDataChanged` already
takes care of the rest (like in `ClearFunctions`).

Bug: http://b/245680119

Test:
- Add unit test.
- As part of the larger symbol loading fallback prototype.