Add SymbolHelper::FindObjectInCache and VerifyObjectFile
`SymbolHelper::FindObjectInCache` looks for the module itself, so it checks both
the build-id and the size. The corresponding `VerifyObjectFile` is added as
well, which checks for both build id and file size as we want an exact match.

This will be used in the context of the "fallback symbols": when symbols are not
available, we cache the module itself. This method will check whether the module
is already in the cache (it doesn't need to have symbols) after the whole search
for symbols has completed (including in the cache).

Bug: http://b/245680119

Test:
- New unit tests.
- As part of the larger symbol loading fallback prototype.