Add ModuleData::SymbolCompleteness and ModuleData::AddFallbackSymbols (#4214)
In terms of loaded symbols, previously a `ModuleData` could be in two states:
symbols not loaded or symbols loaded (`ModuleData::is_loaded()`).

In the context of the "fallback symbols" we will have three states, represented
by the new `ModuleData::SymbolCompleteness`: symbols not loaded, fallback
symbols loaded, full debug symbols loaded. `ModuleData::is_loaded()` is replaced
by `GetLoadedSymbolsCompleteness()`. We also add the convenience methods
`AreDebugSymbolsLoaded()` and `AreAtLeastFallbackSymbolsLoaded()`: in replacing
the usages of `is_loaded()`, in some cases we want to use one, in some cases the
other.

`ModuleData::AddFallbackSymbols` is added to actually add the "fallback
symbols". But note that it is not yet used in this change, only in tests.

Bug: http://b/245680119

Test:
- Extend various unit tests.
- As part of the larger symbol loading fallback prototype.