Only copy stack relevant data if needed
Collecting callstacks on thread state tracepoints introduces
quite some overhead. Finally, we are only interested in
callstacks that belong to our process. So we can save
us from copying that data out of the ring buffer.

Profiling OrbitService showed, that copying the callstack
data was (and still is) the biggest chunk of work
when profiling with callstack on thread state
tracepoints.

Bug: http://b/245529464
Test: Test E2E prototype