Make stack size for thread state slice configurable (on service)
Copying stacks for callstacks on thread state slices is the most
expensive part on OrbitService. For thread state slices,
we don't necesarrily need a complete callstack. We are mostly
interested in the 1-3 callers of the lock/unlock calls.

This is different for normal stack sampling, where we need the
full callstack for our statistics. Thus, it is desirable to
allow the user to specify a different stack size to copy
on thread state slices and reduce the overhead, both
on the game and on OrbitService.

This change is adding a new field in the capture options
to specify the stack dump size for thread states, and
passes it down to the perf_event_open calls. The new
default is 256.

We also streamline some parameters that were passed
instead of using the class member and make stack size
0 if not needed (only for consistency).

We also fixed a typo in the capture options proto.

Test: Profile E2E prototype.
Bug: http://b/243510345