Create ScopeCollection for stat groupings
This new class keeps track of a set of scopes and their relative
stats/ordered timer durations together.

Currently, it's only being used to keep track of all scopes in
CaptureData, but in a future PR it will be used to calculate the stats
of a time range selection of scopes.