Allow to customize the function matching in Mizar (#4224)
This introduces a concept of FunctionSymbolToKey. The class maps function 
symbols to some keys, the functions with equal keys are matched. The function
AssignSampledFunctionIds is wrapped into `BaselineAndComparisonHelperTmpl`
class. The class takes FunctionSymbolToKey as a template parameter. This 
customisation is necessary for unit-testing, where FunctionSymbolToKey is
mocked. As a consequence of templatizing we have to move 
AssignSampledFunctionIds to the header.

Test: Compile, Unit, Manual
Bug: http://b/247977204