Fix Coff file build-id generation (#4234)
It seems we need to shuffle the first 8 of the 16 guid bytes we read in Coff files
in order to generate build ids that match those we get from dumpbin.exe and
that are expected to query the Microsoft Symbol Server. Documentation on the
subject seems non-existent, but it is mentioned in the "CodeView" section in
http://www.godevtool.com/Other/pdb.htm.

With the change, we can now retrieve symbols for the Windows libraries and the
build-ids we generate match the ones from dumpbin.exe.