[Draft] Fix manual instrumentation in the context of pid namespaces.
Previously pids/tids from events obtained from our manual
instrumentation api contained the pid/tid form the namespace in which
the target process is running.
With this PR we translate these pids/tids into the root namespace
before sending them to the client.

This requires a knowledge of the tid mapping in ProducerEventProcessor.
At the beginning of the capture a snapshot containing the data for all
existing threads is send. During the capture the mapping is updated as
threads are created.

Since the arrivial of manual instrumentation events and the update of
the tid mapping are asyncronous we need to buffer manual instrumentation
events that are not (yet) translatable. As threads become known to the
tid mapping we clear out the buffers. The order of manual
instrumentation events from a specific thread is unchanged.

Finally, the introspection events need to bypass the above mechanism:
the tids from the root namespace.