CppWin32 Fixes
Expose variables, delete unused code and fix bugs in
third_party/cppwin32. This makes it possible to use
cppwin32 as a dependency in order to parse the
Windows.Win32.winmd Windows Api file, which uses the
ecma-335 file format, and generate code.

Those changes were tested end-to-end
in the Windows Api Shim branch.