Remove uses of CaptureData::instrumented_functions
This removes all uses of instrumented_functions_ outside of CaptureData.
As this PR has already become really big, removing it entirely will come
in a follow up PR that can correctly move and test the logic of
ComputeVirtualAddressOfInstrumentedFunctionsIfNecessary into
ScopeIdProvider.

This also adds FindFunctionIdSlow to ScopeIdProvider to reduce the
changes in this PR. Some of the uses of this function end up going from
function_id->function_info->function_id and would be better to refactor
so transformations like this don't happen, but this should be done in a
follow up PR.

Bug: http://b/249262736