Use ScopeStatsCollection for LiveTab
This PR sets ScopeStatsCollection to a new collection every time the
user selects a new time range. This will update the live tab such that
only functions in the time range are shown and the stats are updated
accordingly.

This feature is still hidden behind the flag --time_range_selection

Bug: http://b/240112049