Prevent E2E test failure after integration of symbol loading fallback
This is in preparation of merging https://github.com/google/orbit/pull/4241.
The interesting part is that we can't expect no symbol to be in the function
list if debug symbols are not available, because fallback symbols can be loaded
instead.

Note that this is not actually testing the new feature yet.

Bug: http://b/245680119

Test: Run orbit_symbol_location.py locally, with a version of Orbit from
https://github.com/google/orbit/pull/4241.