Integrate symbol loading fallback into symbol loading logic in OrbitApp (#4241)
Primarily, it adds the fallback logic to
`OrbitApp::RetrieveModuleAndLoadSymbols`.
This makes use of the two new `RetrieveModuleSymbolsAndLoadSymbols` and
`RetrieveModuleItselfAndLoadFallbackSymbols`.
And in turns, `RetrieveModuleItselfAndLoadFallbackSymbols` calls the two new
`RetrieveModuleItself` (which also uses `RetrieveModuleItselfFromInstance`) and
`AddFallbackSymbols`..

The formatting of some error messages is improved, in particular nested indented
ones.

Some method declarations are shuffled around in `App.h` for better grouping of
related ones. Also, `RetrieveModule` is renamed to `RetrieveModuleSymbols`.

Finally, `/bigobj` is enabled also for non-debug builds as `App.cpp` got even
larger.

Bug: http://b/245680119

Test: Manual testing with automatic symbol loading and manual reloading in
on modules with various symbol availability and on various scenarios of
connection to the instance.