Remove conan lockfiles
Lockfiles have a couple of bugs which for example prevent us from
updating gRPC. We also don't really need them anymore because their main
purpose was to ensure we have the exact version of the dependency in
releases.

This commit also updates the script that we use to build packages. I
needed to rebuild some packages of the public build to keep things
running without the lockfiles.