Use imported copy of Abseil
This replaces the usage of Abseil from Conan. The main reason for doing
this is that we wanna provide a build that works without Conan (for
building Linux packages). But Abseil packages in most Linux
distributions are compiled in C++11 mode while we need C++17 mode. So
for now we will add our own copy of Abseil to the build.