Update vulkan dependencies
This updates all Vulkan dependencies (headers, validation layers,
loader) to the same version which is 1.3.224.1. An update was inevitable
because the validation layer package is not available for the old
version of Vulkan headers that we used to use.

The given version is the latest version available in Conan. It would
also be possible to downgrade to some earlier version - maybe Vulkan
1.2, but without having any requirements I don't think that makes sense.
Most Linux distributions come with some 1.3 version of Vulkan, so we
need to support 1.3 anyway.

So there are the following things happening:

1. Update Vulkan Header to 1.3.224.1
2. Require the Vulkan Validation Layers Conan package (same version) and
   remove the copy of the header from third_party/
3. Adjustment of the version of Volk (the loader)
4. Remove usage of deprecated macros from the vulkan tutorial code. (I
   made some tests. The code seems still to run - no crash.)