Optimize Callstack Thread Bar by only rendering one per pixel
In this PR we are following the plan designed in
go/stadia-orbit-rendering-60fps for CallstackThreadBar.

We are basically going through pixels instead of all the callstacks. The
new methods were renamed by adding "Discretized", I'm not sure if this
is the best name but I wanted to be consistent with the optimization in
SchedulerTrack and ThreadTrack.

Just because these modified lines, in the 5-min capture presented in the
design doc we are archieving around 2.5x improvement overall (measured in
Linux, from 750ms to 300ms):
Before: http://screenshot/7edKS46zphJVxAL
After: http://screenshot/3qsQLTqLqZXUjgs

Bug: http://b/253414641.
Test: Load 5 min capture.