Support gRPC and protobuf as system dependencies
This also makes a change to user space instrumentation.

When using gRPC and protobuf from system dependencies we are required to
link dynamically. This messes with our tests because we can't inject the
same proto definitions twice into the same process when both proto
definitions share the same protobuf implementation (which is the case
for dynamic linking of libprotobuf).

To solve that this change reworks the library injection code to use
dlmopen() instead of dlopen and open a new library namespace in the
target process. That means libprotobuf can be loaded twice and both
set of symbols won't interfere. This also helps encapsulating the
instrumentation library better from the target process.