More conanfile improvements (#4291)
* Remove OrbitService debian packaging

This will not be needed anymore for a pure open source build.

* Do not run python tests from conan anymore.

This simplifies the conan file quite a bit and simplifies the build
process. We can bring this back later if we think this is important.

* Remove Chromium License Gathering

This is not needed anymore since we are not using Chromium in Orbit.