Avoid dangling reference with old version of outcome (#4313)
I'm trying to make the code base compatible with old versions of OUTCOME
where the OUTCOME_TRY macro was capturing the expression result in a
universal reference instead of in a value.

I.e. this
```cpp
auto&& result = (expr);
if (result.has_value()) out = std::move(result.value());
```

instead of this:
```cpp
auto result = (expr);
if (result.has_value()) out = std::move(result).value();
```

This leads to a compile error (dangling reference), when `expr` is
returning a reference to a temporary with a shorter lifetime.

This is the case for `LiftApply` because it retruns a `Typedef`
containing an `ErrorMessageOr`. The temporary Typedef will be destroyed
before `result` (from the example above) gets out of scope.

A fix is to avoid `OUTCOME_TRY` in this case. It's the only location
in the code that suffers from this problem.