Introduce with_system_deps Conan option
This option will allow to build Orbit (from Conan) but linked against
system dependencies for all dependencies that support that. So far that
is none of the Conan dependencies, but they will all be added one after
the other in subsequent PRs.

This PR also sets up the clang7 presubmit job as a "with system
dependencies" job such that we can test both build modes with the CI.