Create tests and optimize rendering of AllThreadsCallstackBar (#4296)
* Create tests and optimize rendering of AllThreadsCallstackBar

In this PR we are optimizing the rendering of AllThreadsCallstackBar.

In the previous PR (https://github.com/google/orbit/pull/4271), we skip going through and rendering events from the same pixel in a particular thread_id but we were still going once per pixel per thread_id. In this PR we are still iterating through pixels but stopping looking for new callstacks when we have already found some on the current_pixel.

For getting the pixel number I created a method in FastRenderingUtils class, but I also used in CallstackThreadBar since is way more precise. Ideally we should fix these functions (http://b/242973688), but I won't work on that for now. It also makes sense to use GetPixelNumber when we want to render vertical lines.

I found this issue when creating tests for CallstackData (there were no tests before for this method but I should have created them anyway), so I'm including the tests here as well.

Rendering all_thread Track for a 5 min-capture takes:
- Before (https://github.com/google/orbit/pull/4271) - >150ms
- After (https://github.com/google/orbit/pull/4271) - ~13ms
- Now - ~0.950ms

Rendering callstacks for all Tracks takes 8.6% instead of 11.9% (and before https://github.com/google/orbit/pull/4271 - 66%) . I think that this is enough for our goal.

Test: Load a capture, rendering looks fine. Run Unit tests.
Bugs: http://b/252998254 (GetPixelNumber) & http://b/253414641 (Callstack optimization itself)

* Fixup: Davide's comments

* Fixup: Big refactor based on comments

* Fixup: kAllEvents

* Fixup: Follow C++ recommendations