Create a method to consistently draw boxes in OrbitGl (#4332)
* Create a method to consistently draw boxes in OrbitGl

In this PR we are just moving a method to draw boxes from SchedulerTrack
to TimelineInfoInterface. We are having glitches when drawing boxes in
the Capture Window (for example, see https://github.com/google/orbit/pull/4127 for ThreadTracks).
The implementation of this method will extend the box to the pixel
borders as it's doing in the SchedulerTrack. This will be needed for
ThreadStates (https://github.com/google/orbit/pull/4274), to draw consistently after applying
the pixel rendering optimization.

Test: Load a capture, everything looks as expected.

* Fixup: Empty line