Optimize rendering of ThreadStates
In this PR we are following the plan designed in go/stadia-orbit-rendering-60fps for ThreadStates. We are optimizing its rendering by iterating through pixels as we recently implemented for both SchedulerTrack (https://github.com/google/orbit/pull/4157) and CallstackThreadBar (https://github.com/google/orbit/pull/4271).

The implementation is quite similar to CallstackThreadBar:
- ForEach...Discretized method in CaptureData who iterates through pixels using lower_bound queries.
- UnitTests in CaptureDataTest.
- A completely simplified implementation in ThreadStateBar that use the new method instead of having its own implementation

As a result we are rendering each ThreadStateBar around 4x times faster when zooming-out in the 5-min capture introduced in the design doc (1.1-1.4ms instead of 4.4-4.9ms for the one with more events).

The result can be seen here: http://screenshot/6iLSpkSr5QJc3Ka

We are also fixing an issue. Before this PR blue slices has bigger priority to be rendered and were more visible misleading the user (Border case: http://screenshot/5CNdZd89H5wn9pi). Now we are taking the first one at each pixel. This isn't perfect (because anti-aliasing), but it's still better than the current implementation.

A complete analysis will be done comparing a old version with the current one, but as we are about 4x in ThreadState rendering and ThreadState rendering was about 13.9% of the total rendering, I expect an overall ~10.4% improvement in Rendering just by this PR when zooming-out in the 5-min capture.

Bug: http://b/253429871
Test: Load a capture. Zoom-out entirely, zoom-in to the maximum. See ThreadStates rendering looks as it's expected.