Make producer side server's listening socket a flag
The purpose of this change was to make the producer side channel socket
path changeable in the OrbitServiceIntegrationTest to avoid test
failures when in parallel an instance of OrbitService was running on the
same machine.

For consistency reasons I also now made the socket path a flag for
OrbitService since this parameter was already wired through all the way.

I also did some cleanup on the way (Linux/Posix API to STL, std::string
to std::string_view, etc.).