Use Qt text rendering in capture window.
Previously we used freetype and opengl to render the text directly. This
PR changes this. We now use QPainter to draw the text.
In OrbitGLWidget::paintGL we obtain a QPainter and we hand it down to
all the rendering routines.
In GlCanvas::Render we call beginNativePainting and endNativePainting on
the QPainter. So we are - as previously - in native painting mode almost
all the time. Only in CaptureWindow::RenderAllLayers, when we draw the
text layers, we interrupt the native painting and utilize QPainter
directly.
The actual text rendering is done in QtTextRenderer. It replaces the old
OpenGlTextRenderer. QtTextRenderer is not behaving identically to the
old OpenGlTextRenderer. It merely matches the visual result.
Specifically there are differences in the way width and height of the
text is measured; this was a very manual process previously and is now
done by Qt.

There is no implementation of RenderDebug in QtTextRenderer. RenderDebug
looks as if it was used to debug the low level text rendering
previously. I'll remove RenderDebug in a follow up PR.