Use ConnectToLocalWidget in SessionSetupDialog
This makes use of the ConnectToLocalWidget in SessionSetupDialog. The
general workings of SessionSetupDialog are kept for now, but can be
simplified in the future, since all parts (local, stadia and file) are
now its own widget.

Note: This also removes the UI code from ProcessLauncherWidget. This
was Windows only functionality and can be added again to
ConnectToLocalWidget when the Windows build is working.