Local profiling widget + Start OrbitService (#4380)
* Add OrbitServiceInstance to Connections

and use it in orbitmainwindow.

* Add ConnectToLocalWidget

This adds a widget for connecting to OrbitService and starting
OrbitService (if its not done manually). The ConnectToLocalWidgetTest
tests some functionality, but is more of a smoke test, since the grpc
connections can't be mocked currently.

* Use ConnectToLocalWidget in SessionSetupDialog

This makes use of the ConnectToLocalWidget in SessionSetupDialog. The
general workings of SessionSetupDialog are kept for now, but can be
simplified in the future, since all parts (local, stadia and file) are
now its own widget.

Note: This also removes the UI code from ProcessLauncherWidget. This
was Windows only functionality and can be added again to
ConnectToLocalWidget when the Windows build is working.