Do not require cgroup information for memory tracing
Currently, we are only supporting cgroup version 1. On systems
using cgroup version 2, we currently won't produce memory
information at all, as the process and system memory
events keep waiting for the cgroup info to appear.

A quick fix for this is to disable the requirement
if cgroup infor can't be parsed.

Follow-up: This change does not fix the same issue
on page-fault information.

Test: Take a capture with memory tracing and cgroup 2.
Bug: #4408