Make TimeGraphLayout a parameter to TimeGraph (#4416)
The ImGuiTimeGraphLayout used to live in TimeGraph (as a value). This
change makes the TimeGraphLayout a non-owning dependency of TimeGraph -
passed in as a constructor argument and hence following dependency
injection.

Eventually we want a Qt widget (child element of OrbitMainWindow)
provide an implementation of TimeGraphLayout, so we need to pass down
the dependency up the hierarchy - meaning:

1. TimeGraph gets the layout from CaptureWindow
2. CaptureWindow gets the layout from its factory function
   GlCanvas::Create
3. GlCanvas::Create gets the layout from OrbitGLWIdget::Initialize
4. OrbitGLWidget::Initialize is called from OrbitMainWindow, so all is
   good.

Temporarily OrbitMainWindow passes in instances of the
StaticTimeGraphLayout.