Fix GraphTrack arrow boxes being too height
In this PR we are fixing arrow boxes in GraphTrack. After investigating,
seems to be that before we were calculating the height of the text
manually (because GetStringHeight() didn't count multi-lines).

Since now GetStringHeight() is returning the correct height, the visible
height was the square, and then the box was incredible big.

New Screenshot: http://screenshot/4B4n4ZggDKau4Ga